% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_mu.R
\name{arima_mu}
\alias{arima_mu}
\title{Capture \code{arima.mu} from x13 output}
\usage{
arima_mu(sa)
}
\arguments{
\item{sa}{A \code{\link{x13}} output object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Capture \code{arima.mu} from x13 output
}
\examples{
myseries <- pickmdl_data("myseries")

a <- x13(myseries, spec = "RSA3")
b <- x13(sqrt(myseries), spec = "RSA3")

arima_mu(a)
arima_mu(b)

}
