% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corona_outliers.R
\name{corona_outliers}
\alias{corona_outliers}
\alias{update_spec_corona_outliers}
\title{Corona outliers}
\usage{
corona_outliers(option = "ssb", freq = 12, day = "01", q_month = 1)

update_spec_corona_outliers(
  spec,
  option = "ssb",
  freq = 12,
  day = "01",
  q_month = 1
)
}
\arguments{
\item{option}{Only \code{"ssb"} implemented}

\item{freq}{frequency, \code{4} or \code{12}}

\item{day}{day of month as character}

\item{q_month}{month of quarter as \code{1}, \code{2} or \code{3}}

\item{spec}{An \code{\link{x13_spec}} output object to be updated}
}
\value{
data frame
}
\description{
Corona outliers as a type-date data frame. Also a function to update spec with these outliers.
}
\details{
Corona outliers with same date as outliers already in spec will be omitted.
}
\examples{
corona_outliers()
corona_outliers(freq = 4)

spec_a <- x13_spec(spec = "RSA3", transform.function = "Log")
spec_a2 <- update_spec_corona_outliers(spec_a)
s_preOut(spec_a)
s_preOut(spec_a2)

spec_b <- x13_spec(spec = "RSA3", transform.function = "Log", 
                   usrdef.outliersEnabled = TRUE, 
                   usrdef.outliersType = rep("AO", 3), 
                   usrdef.outliersDate = c("2009-01-01", "2016-01-01", "2020-05-01"))
spec_b2 <- update_spec_corona_outliers(spec_b)
s_preOut(spec_b)
s_preOut(spec_b2)
}
\seealso{
\code{\link{s_preOut}}
}
