% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crit_selection.R
\name{crit_selection}
\alias{crit_selection}
\title{Selection from table of PICKMDL criteria}
\usage{
crit_selection(crit_tab)
}
\arguments{
\item{crit_tab}{Output from  \code{\link{crit_table}}}
}
\value{
Selected index
}
\description{
Selection from table of PICKMDL criteria
}
\examples{
myseries <- pickmdl_data("myseries")

spec5 <- x13_spec_pickmdl(spec = "RSA3", transform.function = "Log")

sa5 <- x13_multi(myseries, spec = spec5)   

tab <- crit_table(sa5)

crit_selection(tab)
crit_selection(tab[2:5, ])
crit_selection(tab[1:2, ]) # Warning

}
