% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crit_table.R
\name{crit_table}
\alias{crit_table}
\alias{crit123_m_aic}
\alias{crit1}
\alias{crit2}
\alias{crit3}
\alias{m_aic}
\title{Table of PICKMDL criteria}
\usage{
crit_table(sa_list)

crit123_m_aic(sa)

crit1(sa)

crit2(sa)

crit3(sa)

m_aic(sa)
}
\arguments{
\item{sa_list}{List of several \code{x13} output objects. That is, \code{spec} can be output from  \code{\link{x13_multi}}.}

\item{sa}{A single \code{x13} output object.}
}
\value{
A matrix, a vector or a single numerical value
}
\description{
Function \code{crit_table} takes several \code{x13} output objects as input and produces a table of criteria (matrix class).
The other functions are underlying functions that take a single \code{x13} output object as input.
}
\examples{
myseries <- pickmdl_data("myseries")

spec5 <- x13_spec_pickmdl(spec = "RSA3", transform.function = "Log")

sa5 <- x13_multi(myseries, spec = spec5)   

crit_table(sa5)

crit123_m_aic(sa5[[4]])

crit1(sa5[[4]])
crit2(sa5[[4]])
crit3(sa5[[4]])
m_aic(sa5[[4]]) 
}
