% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalfak.R
\name{konstruksjon}
\alias{konstruksjon}
\title{Lage faktorer for kalendereffekter}
\usage{
konstruksjon(
  forste_ar,
  siste_ar,
  k_roddag = TRUE,
  t_roddag = "sondag",
  rod_dag = c("jan1", "mai1", "mai17", "des25", "des26"),
  k_fpk = TRUE,
  ant_fpk = 3,
  t_fpk = "egen_effekt",
  paske_egen = 1,
  paske_mdl = "X12",
  t_pklordag = "sondag",
  t_pinse = "sondag",
  k_td = FALSE,
  td_type = "TD6",
  k_grupper = FALSE,
  monster = c(1, 2, 3, 4, 5, 6),
  fjerne_se_td = TRUE,
  skudd_ar = TRUE
)
}
\arguments{
\item{forste_ar}{kalenderfil lages fra januar forste aar}

\item{siste_ar}{kalenderfil lages til og med desember siste aar.
Siste ar kan settes langt fram i tid, og boer settes minst ett aar lenger
enn dataene kalenderfilen skal benyttes paa.}

\item{k_roddag}{kan settes til enten TRUE eller FALSE.
Dersom den er satt til TRUE og parameteren
't_roddag' er satt til 'egen_effekt', vil korrigerte faktorer genereres for de
roedene dagene som er spesifisert i parameteren 'rod_dag'. Hvis 't_roddag'
er satt til 'sondag', vil de spesifiserte roededagene bli behandlet som ekstra
soendager. Dersom parameteren 'k_roddag' er satt til FALSE,
er det ikke noedvendig aa spesifisere verdier for parameterne
't_roddag' og 'rod_dag'. Default verdien er TRUE.}

\item{t_roddag}{to valgmuligheter 'egen_effekt' eller 'sondag'.
Hvis parameteren k_roddag er TRUE og t_roddag er 'egen_effekt',
vil effekter av roede dager bli estimert med separate variable,
ellers vil de bli betraktet som ekstra soendager i TD/WD variable.
Default verdien er 'sondag'.}

\item{rod_dag}{inneholder de faste hellig- og hoeytidsdagene (roede)
som skal korrigeres, evt. andre dager (for eksempel des24 og des31).
Kristihimmelfartsdag blir automatisk behandlet som en ekstra soendag i programmet
naar TD-/WD-variable konstrueres}

\item{k_fpk}{TRUE eller FALSE. Hvis det blir TRUE, vil variable for en periode
foer skjaertorsdag bli konstruert. Default verdien er TRUE. Dersom parameteren
k_fpk er satt til FALSE, er det ikke noedvendig aa spesifisere verdier
for ant_fpk og t_fpk.}

\item{ant_fpk}{tallet representerer antall dager foer skjaertorsdag som
blir justert paa grunn av endret aktivitet i tilknytning til paasken (som kan havne
i mars eller april). For eksempel vil ant_fpk lik 3 bety mandag, tirsdag og onsdag
foer skjaertorsdag. Default verdien er 3.}

\item{t_fpk}{to valgmuligheter 'egen_effekt' og 'sondag'.
Hvis det blir 'egen_effekt' vil effekten foer paasken bli estimert ellers
det blir betrakt som ekstra soendager. Default verdien er 'egen_effekt'.}

\item{paske_egen}{kan settes til enten 1 eller 2. Hvis det settes til 1 (default), vil
Skjaertorsdag, Langfredag og 2. paaskedag regnes som ekstra soendager. Paaskeaften
kan enten regnes som en soendag eller ha sin egen effekt. Hvis paaske_egen derimot settes
til 2, vil Skjaertorsdag, Langfredag, paaskeaften og 2. paaskedag ha sin egen felles effekt.
Paaskevariablen er sesongrenset.}

\item{paske_mdl}{'X12' eller 'linear_reg'. Hvis vi antar at effekten er jevn den samme av
alle prepaaske dagene, vil metoden 'X12' bli brukt,
ellers vil metoden 'linear_reg' bli brukt, som betyr lineeart stigende effekt av hver dag
i paaske perioden eller prepaaske perioden. Default verdien er 'X12'.}

\item{t_pklordag}{'egen_effekt' eller 'sondag'. Hvis 'egen_effekt' velges,
vil effekten av paaske-aften bli estimert, ellers vil det bli betrakt som en soendag.
Default verdien er 'sondag'.}

\item{t_pinse}{'sondag' eller 'egen_effekt'. Hvis 'egen_effekt' velges, vil effekten av 2.
pinse-dag bli estimert, ellers vil det bli betrakt som en soendag. Default verdien er 'sondag'.}

\item{k_td}{TRUE eller FALSE. Hvis det blir TRUE, vil trading day variable bli laget.
Default verdien er TRUE. Dersom parameteren 'k_td' er satt til FALSE, er det ikke
noedvendig aa spesifisere verdier for 'td_type', 'k_grupper' og 'monster'.}

\item{td_type}{Det er 4 valgt TD5, TD6, TD15 og TD16. Der 'TD5' har 5 arbeidsdager i uken
fra mandag til fredag med ulike aktiviteter. Loerdag og soendag er fri.
'TD6' har 6 handledager i uken fra mandag til loerdag, soendag er fri.
'TD15' er en variant av 'TD5' hvor det antas samme aktivitet
paa de 5 arbeidsdagene fra mandag til fredag. 'TD16' er en variant av
'TD6' hvor handledagene har de samme handle intensitet. Maa sette k_td = TRUE hvis du vil at programmet kjoeres med verdien av td_type.}

\item{k_grupper}{TRUE eller FALSE. Hvis det er TRUE, vil ukedagene bli gruppert i grupper.
Default verdien er FALSE. Dersom parameteren 'k_grupper' er FALSE, er det ikke noedvendig
aa spesifisere verdi til 'monster'.}

\item{monster}{Dette viser hvilke ukedager som vil bli gruppert sammen,
der 6 elementer spesifiseres dersom TD6 benyttes og 5 elementer spesifiseres dersom
TD5 benyttes over. For eksempel, hvis 'monster' er satt til c(1, 1, 1, 1, 2, 2),
vil mandag, tirsdag, onsdag og torsdag bli gruppert sammen i en gruppe,
mens fredag og loerdag vil vaere i en annen gruppe. Ett annet eksempel er
c(1, 2, 2, 2, 3), da mandag en egen gruppe, mens  tirsdag, onsdag og torsdag er
en annen gruppe, mens fredag er en 3. gruppe. Da er loerdag og soendag kontrast gruppen.}

\item{fjerne_se_td}{TRUE eller FALSE. Hvis det blir TRUE, vil sesong av kalendereffekter
bli trukket fra trading day regressorer av. Default verdien er TRUE. Hvis TRUE
fjernes sesongeffektene i TD variable basert paa perioden angitt (i forste_ar og siste_ar).}

\item{skudd_ar}{TRUE eller FALSE. Hvis det settes til TRUE vil programmet lage en variabel "skudd_ar" for skuddaar.
Default verdien er TRUE,}
}
\value{
korrigerte faktorer
}
\description{
Dette programmet genererer korrigerte faktorer for kalendereffekter
}
\note{
This function is copied from an internal package (kalfak) at SSB maintained by Dinh Pham.
}
\examples{
konstruksjon(forste_ar = 2000, siste_ar = 2050, 
  k_roddag = TRUE, t_roddag = "sondag", 
  rod_dag = c("jan1", "mai1", "mai17", "des24", "des25", "des26"), 
  k_fpk = TRUE, ant_fpk = 7, t_fpk = "egen_effekt", 
  paske_egen = 1, paske_mdl = "X12", t_pklordag = "egen_effekt",
  t_pinse = "sondag", k_td = TRUE, td_type = "TD6", 
  k_grupper = TRUE, monster = c(1, 1, 2, 2, 2, 3), 
  fjerne_se_td = TRUE, skudd_ar = TRUE)

}
