% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ok.R
\name{ok}
\alias{ok}
\title{PICKMDL information as a list}
\usage{
ok(sa)
}
\arguments{
\item{sa}{Output from \code{\link{x13_pickmdl}}}
}
\value{
List constructed from comment attribute
}
\description{
Extract \code{ok}, \code{ok_final} and \code{mdl_nr}
}
\examples{
myseries <- pickmdl_data("myseries")

a <- x13_pickmdl(myseries, x13_spec(spec = "RSA3", transform.function = "Log"))
b <- x13_pickmdl(myseries, x13_spec(spec = "RSA3", transform.function = "None"))

comment(a)
comment(b)
ok(a)
ok(b)

}
\seealso{
\code{\link{crit_ok}}
}
