% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickmdl_data.R
\name{pickmdl_data}
\alias{pickmdl_data}
\title{Function that returns an r object with data}
\usage{
pickmdl_data(dataset)
}
\arguments{
\item{dataset}{Name of data set within the pickmdl package}
}
\value{
An r object with data
}
\description{
Function that returns an r object with data
}
\details{
\strong{myseries:} A univariate time series made from \code{pickmdl_data("datainn")}

\strong{allvar:} Matrix of time series (mts) made from \code{pickmdl_data("kalendernr3")}

\strong{datainn:} A vector of data.

\strong{kalendernr3:} As \code{read.table("kalendernr3.txt")}
}
\examples{
pickmdl_data("myseries")
pickmdl_data("qseries")
pickmdl_data("allvar")
pickmdl_data("datainn")
pickmdl_data("kalendernr3")
}
