% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_spec_outliers.R
\name{update_spec_outliers}
\alias{update_spec_outliers}
\alias{update_outliers}
\title{Update x13 spec with outliers}
\usage{
update_spec_outliers(spec, sa, day = "01")

update_outliers(spec, sa, day = "01", null_when_no_new = TRUE)
}
\arguments{
\item{spec}{An \code{\link{x13_spec}} output object}

\item{sa}{An \code{\link{x13}} output object}

\item{day}{Day of month as character to be used in outlier coding}

\item{null_when_no_new}{Whether to return \code{NULL} when no new outliers found.}
}
\value{
\code{update_spec_outliers} returns an updated \code{x13_spec} output object with
new outliers and updated \code{outlier.from}.
\code{update_outliers} returns a data frame with outlier variables used to update.
}
\description{
Update an \code{x13_spec} output object with outliers from an \code{x13} output object.
}
\examples{
myseries <- ipi_c_eu[, "FR"]

spec_1 <- x13_spec(spec = "RSA3", transform.function = "None", usrdef.outliersEnabled = TRUE, 
                   usrdef.outliersType = "AO", usrdef.outliersDate = "2007-03-01", 
                   outlier.usedefcv = FALSE, outlier.cv = 3)
                   
spec_2 <- x13_spec(spec_1, estimate.to = "2018-08-01")

a <- x13(myseries, spec_2)

update_outliers(spec_1, a)

spec_3 <- update_spec_outliers(spec_1, a)

s_span(spec_1)
s_span(spec_2)
s_span(spec_3)

s_preOut(spec_1)
s_preOut(spec_2)
s_preOut(spec_3)
}
