% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x13_pickmdl.R
\name{x13_pickmdl}
\alias{x13_pickmdl}
\title{x13 with PICKMDL}
\usage{
x13_pickmdl(
  series,
  userdefined = NULL,
  ...,
  arima.p = c(0, 0, 2, 0, 2),
  arima.d = c(1, 1, 1, 2, 1),
  arima.q = c(1, 2, 0, 2, 2),
  arima.bp = 0,
  arima.bd = 1,
  arima.bq = 1,
  automdl.enabled = FALSE
)
}
\arguments{
\item{series}{\code{x13} parameter}

\item{userdefined}{\code{x13} parameter}

\item{...}{\code{x13_spec} parameters}

\item{arima.p}{\code{x13_spec} parameter as vector}

\item{arima.d}{\code{x13_spec} parameter as vector}

\item{arima.q}{\code{x13_spec} parameter as vector}

\item{arima.bp}{\code{x13_spec} parameter, possibly as vector}

\item{arima.bd}{\code{x13_spec} parameter, possibly as vector}

\item{arima.bq}{\code{x13_spec} parameter, possibly as vector}

\item{automdl.enabled}{\code{x13_spec} parameter}
}
\value{
An \code{x13} output object
}
\description{
\code{\link{x13}} is run with a PICKMDL specification
}
\examples{
myseries <- pickmdl_data("myseries")

a <- x13_pickmdl(myseries, spec = "RSA3", transform.function = "Log")
a$regarima

allvar <- pickmdl_data("allvar")

b <- x13_pickmdl(myseries, spec = "RSA3", transform.function = "Log",
                 usrdef.varEnabled = TRUE, 
                 usrdef.varType = c("Calendar", "Calendar"), 
                 usrdef.var = allvar, 
                 outlier.enabled = FALSE, 
                 usrdef.outliersEnabled = TRUE,
                 usrdef.outliersType = rep("LS", 20), 
                 usrdef.outliersDate = c("2009-01-01", "2016-01-01", 
                                         "2020-03-01", "2020-04-01", "2020-05-01", 
                                         "2020-06-01", "2020-07-01", "2020-08-01", 
                                         "2020-09-01", "2020-10-01", "2020-11-01", 
                                         "2020-12-01", "2021-01-01", "2021-02-01",
                                         "2021-03-01", "2021-04-01", "2021-05-01",
                                         "2021-06-01", "2021-07-01", "2021-08-01"))
b$regarima

# Warning when transform.function = "None"
d <- x13_pickmdl(myseries, spec = "RSA3", transform.function = "None")
                                          
}
