% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x13_text_frame.R
\name{x13_text_frame}
\alias{x13_text_frame}
\title{x13_text_frame}
\usage{
x13_text_frame(
  text_frame,
  series = NULL,
  id = NULL,
  ...,
  drop = TRUE,
  verbose = FALSE,
  dots2list = TRUE
)
}
\arguments{
\item{text_frame}{text_frame}

\item{series}{series}

\item{id}{id}

\item{...}{dots}

\item{drop}{drop}

\item{verbose}{verbose}

\item{dots2list}{dots2list}
}
\description{
x13_text_frame
}
\examples{

myseries <- pickmdl_data("myseries")
seriesABC <- cbind(A = myseries, B = myseries + 10, C = myseries + 20)

tf <- data.frame(name = c("A", "B", "C"), automdl.enabled = c("TRUE", "FALSE", "FALSE"),
                 usrdef.outliersDate = c('c("2009-01-01", "2016-01-01")', 'c("2009-01-01")', NA),
                 usrdef.outliersType = c('rep("LS", 2)', '"AO"', NA),
                 usrdef.outliersEnabled = c("TRUE", "TRUE", NA))

outABC <- x13_text_frame(tf, series = "seriesABC", spec = "RSA3", transform.function = "Log", verbose = TRUE)

}
