% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x13_both.R
\name{x13_both}
\alias{x13_both}
\title{\code{\link{x13_spec}}  and \code{\link{x13_pickmdl}} wrapped as a single function}
\usage{
x13_both(
  series,
  ...,
  userdefined = NULL,
  pickmdl_method = "first",
  star = 1,
  when_star = warning,
  identification_end = NULL,
  identification_estimate.to = NULL,
  identify_t_filter = FALSE,
  identify_s_filter = FALSE,
  identify_outliers = FALSE,
  automdl.enabled = FALSE,
  verbose = FALSE,
  output = "sa",
  both_output = "main"
)
}
\arguments{
\item{series}{\code{x13} parameter}

\item{...}{Parameters to \code{\link{x13_spec}}.}

\item{userdefined}{Parameter to \code{\link{x13}}  (via \code{...}  to \code{x13_pickmdl}).}

\item{pickmdl_method}{\code{\link{crit_selection}} parameter}

\item{star}{\code{\link{crit_selection}} parameter}

\item{when_star}{\code{\link{crit_selection}} parameter}

\item{identification_end}{To shorten the series before runs used to identify (arima) parameters.
That is, the series is shortened by \verb{window(series,} \verb{end = identification_end)}.}

\item{identification_estimate.to}{To set \code{\link{x13_spec}} parameter \code{estimate.to} before runs used to identify (arima) parameters.
This is an alternative to  \code{identification_end}.}

\item{identify_t_filter}{When \code{TRUE}, Henderson trend filter is identified by the shortened (see above) series.}

\item{identify_s_filter}{When \code{TRUE}, Seasonal moving average filter is identified by the shortened series.}

\item{identify_outliers}{When \code{TRUE}, Outliers are identified by the shortened series.}

\item{automdl.enabled}{When \code{TRUE}, automdl is performed instead of pickmdl.
If \code{spec} is a list of several objects as outputted from \code{x13_spec_pickmdl}, only first object is used.}

\item{verbose}{Printing information to console when \code{TRUE}.}

\item{output}{One of \code{"sa"} (default), \code{"spec"} (final spec), \code{"sa_spec"} (both) and \code{"all"}. See examples.}

\item{both_output}{One of \code{"main"} (default, x13_pickmdl output), \code{"spec"} (spec output) or \code{"both"}.}
}
\description{
Output is determined by the parameter \code{both_output}.
}
\details{
All parameters except \code{both_output} and \code{...} are parameters to \code{\link{x13_pickmdl}}.
}
\examples{
a <- x13_both(pickmdl_data("myseries"), spec = "RSA3", transform.function = "Log", verbose = TRUE)
a$decomposition

}
