% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x13_pickmdl.R
\name{x13_pickmdl}
\alias{x13_pickmdl}
\title{x13 with PICKMDL}
\usage{
x13_pickmdl(series, spec, ..., pickmdl_method = "first")
}
\arguments{
\item{series}{\code{x13} parameter}

\item{spec}{List of several \code{x13_spec} output objects. That is, \code{spec} can be output from  \code{\link{x13_spec_pickmdl}}.}

\item{...}{Further \code{x13} parameters (currently only parameter \code{userdefined} is additional parameter to \code{x13}).}

\item{pickmdl_method}{\code{\link{crit_selection}} parameter}
}
\value{
An \code{x13} output object
}
\description{
\code{\link{x13}} is run with a PICKMDL specification
}
\examples{
myseries <- pickmdl_data("myseries")

spec_a  <- x13_spec_pickmdl(spec = "RSA3", transform.function = "Log")

a <- x13_pickmdl(myseries, spec_a)
a$regarima

allvar <- pickmdl_data("allvar")

spec_b <- x13_spec_pickmdl(
            spec = "RSA3", transform.function = "Log",
            usrdef.varEnabled = TRUE, 
            usrdef.varType = c("Calendar", "Calendar"), 
            usrdef.var = allvar, 
            outlier.enabled = FALSE, 
            usrdef.outliersEnabled = TRUE,
            usrdef.outliersType = rep("LS", 20), 
            usrdef.outliersDate = c("2009-01-01", "2016-01-01", 
                                    "2020-03-01", "2020-04-01", "2020-05-01", 
                                    "2020-06-01", "2020-07-01", "2020-08-01", 
                                    "2020-09-01", "2020-10-01", "2020-11-01", 
                                    "2020-12-01", "2021-01-01", "2021-02-01",
                                    "2021-03-01", "2021-04-01", "2021-05-01",
                                    "2021-06-01", "2021-07-01", "2021-08-01"))
b <- x13_pickmdl(myseries, spec_b)                                     
b$regarima

# Warning when transform.function = "None"
spec_d  <- x13_spec_pickmdl(spec = "RSA3", transform.function = "None")
d <- x13_pickmdl(myseries, spec_d) 
                                          
}
