% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x13_pickmdl.R
\name{x13_pickmdl}
\alias{x13_pickmdl}
\alias{x13_automdl}
\title{x13 with PICKMDL and partial concurrent possibilities}
\usage{
x13_pickmdl(
  series,
  spec,
  corona = FALSE,
  ...,
  pickmdl_method = "first",
  star = 1,
  when_star = warning,
  when_automdl = message,
  when_finalnotok = NULL,
  identification_end = NULL,
  identification_estimate.to = NULL,
  identify_t_filter = FALSE,
  identify_s_filter = FALSE,
  identify_outliers = TRUE,
  identify_arima_mu = TRUE,
  automdl.enabled = FALSE,
  fastfirst = TRUE,
  verbose = FALSE,
  output = "sa",
  add_comment = TRUE
)

x13_automdl(..., automdl.enabled = TRUE)
}
\arguments{
\item{series}{\code{x13} parameter}

\item{spec}{An \code{\link{x13_spec}} output object or a list of several objects as outputted from \code{\link{x13_spec_pickmdl}}.
In the case of a single object and when \code{automdl.enabled} is \code{FALSE}, \code{spec} will be converted internally
by \code{x13_spec_pickmdl} with default five arima model specifications.}

\item{corona}{Whether to update \code{spec} by outliers according to \code{\link{corona_outliers}}.
\code{FALSE} or \code{NULL} means no update. \code{TRUE} or \code{"ssb"} means update.}

\item{...}{Further \code{x13} parameters (currently only parameter \code{userdefined} is additional parameter to \code{x13}).}

\item{pickmdl_method}{\code{\link{crit_selection}} parameter
or one of the two extra possibilities, \code{"first_automdl"} or \code{"first_tryautomdl"}.
In both cases the \code{crit_selection} parameter is \code{"first"} and the automdl model is added as the last pickmdl model.
\itemize{
\item \strong{\code{"first_automdl"}:} The automdl model is chosen whenever no pickmdl model is ok.
In other words, the \code{star} parameter changes.
\item \strong{\code{"first_tryautomdl"}:} When no pickmdl model is ok:  The automdl model is chosen if this model is ok,
otherwise the \code{star} model is chosen.
}}

\item{star}{\code{\link{crit_selection}} parameter}

\item{when_star}{\code{\link{crit_selection}} parameter}

\item{when_automdl}{Function to be called when automdl since no pickmdl model ok. Supply NULL to do nothing.}

\item{when_finalnotok}{Function to be called, e.g. \code{\link{warning}}, when final run with final model is not ok. Supply NULL to do nothing.
See \code{\link{crit_ok}}.}

\item{identification_end}{To shorten the series before runs used to identify (arima) parameters.
That is, the series is shortened by \verb{window(series,} \verb{end = identification_end)}.}

\item{identification_estimate.to}{To set \code{\link{x13_spec}} parameter \code{estimate.to} before runs used to identify (arima) parameters.
This is an alternative to  \code{identification_end}.}

\item{identify_t_filter}{When \code{TRUE}, Henderson trend filter is identified by the shortened (see above) series.}

\item{identify_s_filter}{When \code{TRUE}, Seasonal moving average filter is identified by the shortened series.}

\item{identify_outliers}{When \code{TRUE}, Outliers are identified by the shortened series.}

\item{identify_arima_mu}{When \code{TRUE}, \code{arima.mu} is identified by the shortened series (see \code{\link{arima_mu}}).}

\item{automdl.enabled}{When \code{TRUE}, automdl is performed instead of pickmdl.
If \code{spec} is a list of several objects as outputted from \code{x13_spec_pickmdl}, only first object is used.}

\item{fastfirst}{When \code{TRUE} and when pickmdl with \code{crit_selection} parameter \code{"first"},
only as many models as needed are run.
This affects the output when \code{output = "all"}.}

\item{verbose}{Printing information to console when \code{TRUE}.}

\item{output}{One of \code{"sa"} (default), \code{"spec"} (final spec), \code{"sa_spec"} (both) and \code{"all"}. See examples.}

\item{add_comment}{When \code{TRUE}, a  comment attribute
(character vector with \code{ok}, \code{ok_final} and \code{mdl_nr}) will
be added to the \code{\link{x13}} output object. Use \code{\link{comment}}
to get the attribute or \code{\link{ok}} to get the attribute converted to a list.}
}
\value{
By default an \code{x13} output object, or otherwise a list as specified by parameter \code{output}.
}
\description{
\code{\link{x13}} can be run as usual (automdl) or with a PICKMDL specification.
The ARIMA model, outliers and filters can be identified at a certain date and then held fixed (with a new outlier-span).
}
\examples{
myseries <- pickmdl_data("myseries")

spec_a  <- x13_spec(spec = "RSA3", transform.function = "Log")

a <- x13_pickmdl(myseries, spec_a, verbose = TRUE)
comment(a)
ok(a)
unlist(ok(a))
a$regarima

a2 <- x13_pickmdl(myseries, spec_a, identification_end = c(2014, 2))
a2$regarima

# As above, another way
a3 <- x13_pickmdl(myseries, spec_a, identification_estimate.to = "2014-03-01")
a3$regarima

a4 <- x13_automdl(myseries, spec_a, identification_end = c(2014, 2))
a4$regarima

# As above, another way
spec_a_single  <- x13_spec(spec = "RSA3", transform.function = "Log")
a5 <- x13_automdl(myseries, spec_a_single, identification_estimate.to = "2014-03-01")
a5$regarima

allvar <- pickmdl_data("allvar")

spec_b <- x13_spec(
            spec = "RSA3", transform.function = "Log",
            usrdef.varEnabled = TRUE, 
            usrdef.varType = c("Calendar", "Calendar"), 
            usrdef.var = allvar, 
            outlier.enabled = FALSE, 
            usrdef.outliersEnabled = TRUE,
            usrdef.outliersType = rep("LS", 20), 
            usrdef.outliersDate = c("2009-01-01", "2016-01-01", 
                                    "2020-03-01", "2020-04-01", "2020-05-01", 
                                    "2020-06-01", "2020-07-01", "2020-08-01", 
                                    "2020-09-01", "2020-10-01", "2020-11-01", 
                                    "2020-12-01", "2021-01-01", "2021-02-01",
                                    "2021-03-01", "2021-04-01", "2021-05-01",
                                    "2021-06-01", "2021-07-01", "2021-08-01"))
b <- x13_pickmdl(myseries, spec_b, identification_end = c(2020, 2))                                     
b$regarima

# automdl instead  
b1 <- x13_automdl(myseries, spec_b, identification_end = c(2020, 2))
b1$regarima

# effect of identify_t_filter and identify_s_filter
set.seed(1)
rndseries <- ts(rep(1:12, 20) + (1 + (1:240)/20) * runif(240) + 0.5 * c(rep(1, 120), (1:120)^2), 
                frequency = 12, start = c(2000, 1))
spec_c <- x13_spec(outlier.enabled = FALSE)               
c1 <- x13_automdl(rndseries, spec_c, identification_end = c(2009, 12))    
c1$decomposition
c2 <- x13_automdl(rndseries, spec_c, identification_end = c(2009, 12), identify_t_filter = TRUE) 
c2$decomposition
c3 <- x13_automdl(rndseries, spec_c, identification_end = c(2009, 12), identify_t_filter = TRUE, 
                  identify_s_filter = TRUE)     
c3$decomposition                       


# Warning when transform.function = "None"
spec_d  <- x13_spec(spec = "RSA3", transform.function = "None")
d <- x13_pickmdl(myseries, spec_d, verbose = TRUE)

# Warning avoided (when_star) and 2nd (star) model selected 
d2 <- x13_pickmdl(myseries, spec_d, star = 2, when_star = NULL, verbose = TRUE)

# automdl since no pickmdl model ok, but still not ok 
d3 <- x13_pickmdl(myseries, spec_d, pickmdl_method = "first_automdl", verbose = TRUE)

# airline model (star) since automdl also not ok 
d4 <- x13_pickmdl(myseries, spec_d, pickmdl_method = "first_tryautomdl", verbose = TRUE,
                  when_finalnotok = warning) # also finalnotok warning

# As a2, with output = "all"
k <- x13_pickmdl(myseries, spec_b, identification_end = c(2010, 2), output = "all",
                 fastfirst = FALSE) # With TRUE only one model in this case 
k$sa$decomposition  # As a2$decomposition 
k$mdl_nr            # index of selected model used to identify parameters
k$sa_mult[[k$mdl_nr]]$decomposition  # decomposition for model to identify
k$crit_tab          # Table of criteria 


# Effect of identify_outliers (TRUE is default)
q1 <- x13_pickmdl(myseries, x13_spec("RSA3", outlier.usedefcv = FALSE, outlier.cv = 3), 
                  identification_end = c(2010, 2), identify_outliers = FALSE)
q2 <- x13_pickmdl(myseries, x13_spec("RSA3", outlier.usedefcv = FALSE, outlier.cv = 3), 
                  identification_end = c(2010, 2), identify_outliers = TRUE, 
                  verbose = TRUE, output = "all")
q3 <- x13_pickmdl(myseries, q2$spec, identification_end = c(2018, 2), identify_outliers = TRUE, 
                  verbose = TRUE)

q1$regarima
q2$sa$regarima
q3$regarima


# With corona outliers (even possible when series is not long enough) 
ac <- x13_pickmdl(myseries, spec_a, verbose = TRUE, corona = TRUE)
ac$regarima
q4 <- x13_pickmdl(myseries, x13_spec("RSA3", outlier.usedefcv = FALSE, outlier.cv = 3), 
                  identification_end = c(2010, 2), identify_outliers = TRUE, 
                  verbose = TRUE, corona = TRUE) 
q4$regarima  
    

}
